`generate_config` now creates paths if needed
If the folder structure passed with outputfile in not existing the
folders are created.

It is also possible to use string substitution in outputfile now.
Default is the lookup in build-vars. But this can be changed with
a different specification in section.

Loading the config dict is a public method now and can be used in
the tasks.py.